+++ src/test/lib/parts_test.ts
@@ -124,6 +124,53 @@ suite('Parts', () => {
         assert.equal(stripExpressionMarkers(container.innerHTML), '');
       });
 
+      test('accepts a symbol', () => {
+        const sym = Symbol();
+        part.setValue(sym);
+        part.commit();
+        assert.equal(stripExpressionMarkers(container.innerHTML), String(sym));
+      });
+
+      test('accepts a symbol with a description', () => {
+        const sym = Symbol('description!');
+        part.setValue(sym);
+        part.commit();
+        assert.equal(stripExpressionMarkers(container.innerHTML), String(sym));
+      });
+
+      test('accepts a symbol on subsequent renders', () => {
+        const sym1 = Symbol();
+        part.setValue(sym1);
+        part.commit();
+        assert.equal(stripExpressionMarkers(container.innerHTML), String(sym1));
+
+        // If the previously rendered value caused a single text node to be
+        // created, then subsequent renders will try to update the existing text
+        // node by setting `.data`. If the new value is a symbol and it isn't
+        // explicitly converted with `String`, then this would throw.
+        const sym2 = Symbol('description!');
+        part.setValue(sym2);
+        part.commit();
+        assert.equal(stripExpressionMarkers(container.innerHTML), String(sym2));
+      });
+
+      test('accepts an object', () => {
+        part.setValue({});
+        part.commit();
+        assert.equal(
+            stripExpressionMarkers(container.innerHTML), '[object Object]');
+      });
+
+      test('accepts an object with a `toString` method', () => {
+        part.setValue({
+          toString() {
+            return 'toString!';
+          }
+        });
+        part.commit();
+        assert.equal(stripExpressionMarkers(container.innerHTML), 'toString!');
+      });
+
       test('accepts a function', () => {
         const f = () => {
           throw new Error();